Add files via upload
Modified the bot to work for a different subreddit. Does not require a formatted title on reddit posts. User may specify how old or new the posts they wish to use are, and the minimum score of the reddit post. Posts are queued to a blog with a title matching the Reddit post title with a link to the URL. Works for queuing photo posts from reddit to tumblr.